+++ cmd/ppa-web/generate.go
@@ -1,4 +1,8 @@
 package main
 
-///go:generate go run github.com/a-h/templ/cmd/templ generate
+import (
+	_ "github.com/a-h/templ"
+)
+
+//go:generate go run github.com/a-h/templ/cmd/templ generate
 //go:generate go fmt ./...